+++ cli/src/pcluster/schemas/cluster_schema.py
@@ -1322,11 +1322,11 @@ class SlurmComputeResourceSchema(_ComputeResourceSchema):
     tags = fields.Nested(
         QueueTagSchema, many=True, metadata={"update_policy": UpdatePolicy.QUEUE_UPDATE_STRATEGY, "update_key": "Key"}
     )
-    static_node_weight = fields.Int(
+    static_node_priority = fields.Int(
         validate=validate.Range(min=1, max=2**32 - 1),
         metadata={"update_policy": UpdatePolicy.SUPPORTED},
     )
-    dynamic_node_weight = fields.Int(
+    dynamic_node_priority = fields.Int(
         validate=validate.Range(min=1, max=2**32 - 1),
         metadata={"update_policy": UpdatePolicy.SUPPORTED},
     )